Add files via upload
Explanation:
URL: We define the URL of the Amazon search results page for laptops. You can modify this URL to scrape different products or from different sections of the website.

get_html function: Sends an HTTP GET request to the URL and returns the HTML content of the page. It uses headers to mimic a request from a web browser.

parse_html function: Uses BeautifulSoup to parse the HTML content and extract product information. It looks for the product name, price, and rating using CSS selectors and stores them in a list of dictionaries.

save_to_csv function: Saves the extracted product information to a CSV file using the pandas library.

Main script: Combines the functions to get the HTML content, parse it, and save the extracted product information to a CSV file.

Note:
Amazon and many other websites have anti-scraping measures in place, and scraping such websites without permission may violate their terms of service. Always check the website's robots.txt file and terms of service before scraping.
The script provided above is for educational purposes. If you need to scrape data for production use, consider using APIs provided by the website, if available.
You can run this script to extract product information from the specified Amazon page and save it to a CSV file named products.csv. Adjust the URL and selectors as needed to target different products or websites.
URL of the website used for web scraping : URL = "http://books.toscrape.com/"